specific test file for namespace restriction
